@@ -16,14 +16,14 @@ classifiers = [
     "Intended Audience :: Developers",
     "License :: OSI Approved :: MIT License",
     "Programming Language :: Python :: 3",
-    "Programming Language :: Python :: 3.8",
     "Programming Language :: Python :: 3.9",
     "Programming Language :: Python :: 3.10",
     "Programming Language :: Python :: 3.11",
     "Programming Language :: Python :: 3.12",
+    "Programming Language :: Python :: 3.13",
     "Topic :: Software Development :: Quality Assurance",
 ]
-requires-python = ">=3.8"
+requires-python = ">=3.9"
 dependencies = [
     "mccabe>=0.7.0",
     "cognitive-complexity>=1.3.0",
@@ -54,7 +54,7 @@ include = [
 ]
 
 [tool.ruff]
-target-version = "py38"
+target-version = "py39"
 line-length = 88
 
 [tool.ruff.lint]
@@ -135,7 +135,7 @@ line-ending = "auto"
 known-first-party = ["pycomplex"]
 
 [tool.mypy]
-python_version = "3.8"
+python_version = "3.9"
 warn_return_any = true
 warn_unused_configs = true
 disallow_untyped_defs = true
@@ -226,4 +226,5 @@ dev = [
     "ruff>=0.12.0",
     "types-tabulate>=0.9.0.20241207",
     "pre-commit>=3.5.0",
+    "tox>=4.0.0",
 ]
